requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
eval_type_backport==0.2.2
faiss-cpu==1.7.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
httpx-sse==0.4.0
hume==0.7.8
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-text-splitters==0.3.6
langsmith==0.3.11
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
numexpr==2.10.2
numpy==1.26.4
openai==1.2.0
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.11.2
simpleaudio==1.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
SQLAlchemy==2.0.38
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websockets==12.0
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0